requirements.txt
telepot
python-chess
pillow
cairosvg
